requirements.txt
django
djangorestframework
drf-spectacular
drf-spectacular[redoc]
pytest
requests
pandas